修改

+++ KinectGestureRecognition/utils.h
@@ -3,8 +3,8 @@
 
 static const int cDepthWidth = 512;
 static const int cDepthHeight = 424;
-static const int cThreshold = 50;
-static const int threshold = cThreshold / 1000.0;
+static const float cThreshold = 50;
+static const float threshold = cThreshold / 1000.0;
 
 template<class Interface>
 inline void SafeRealease(Interface * & pInterfaceToRelease)
